+++ CHANGELOG.md
@@ -6,7 +6,7 @@ CHANGELOG
 - add save-button to individual heroes
 - add buttons to set hero level from experience and vice versa
 - fix game version name not being fully visible (issue #11)
-- fix not being able to modify experience points over 1 billion (2**30)
+- fix not being able to modify experience points over 2^30
 - cap hero level at 75 (74 in Horn of the Abyss)